Pullrequest: Code Cleanup jdk 1.7

+++ app/src/main/java/org/simlar/PreferencesHelper.java
@@ -76,7 +76,7 @@ final class PreferencesHelper
 			digest.update(str.getBytes());
 			final StringBuilder sb = new StringBuilder();
 			for (final byte b : digest.digest()) {
-				sb.append(String.format("%02x", Byte.valueOf(b)));
+				sb.append(String.format("%02x", b));
 			}
 			return sb.toString();
 		} catch (final Exception e) {
