+++ app/src/main/java/org/simlar/ProximityScreenLockerFallback.java
@@ -83,7 +83,7 @@ public class ProximityScreenLockerFallback implements ProximityScreenLocker, Sen
 		final float distance = event.values[0];
 
 		if (distance > event.sensor.getMaximumRange()) {
-			Lg.w(LOGTAG, "proximity sensors distance=", Float.valueOf(distance), " out of range=", Float.valueOf(event.sensor.getMaximumRange()));
+			Lg.w(LOGTAG, "proximity sensors distance=", distance, " out of range=", event.sensor.getMaximumRange());
 			/// do not return, e.g. the galaxy s2 won't work otherwise
 		}
 
@@ -92,14 +92,14 @@ public class ProximityScreenLockerFallback implements ProximityScreenLocker, Sen
 
 		final WindowManager.LayoutParams params = window.getAttributes();
 		if (distance <= PROXIMITY_DISTANCE_THRESHOLD) {
-			Lg.i(LOGTAG, "proximity sensors distance=", Float.valueOf(distance), " below threshold=", Float.valueOf(PROXIMITY_DISTANCE_THRESHOLD),
+			Lg.i(LOGTAG, "proximity sensors distance=", distance, " below threshold=", PROXIMITY_DISTANCE_THRESHOLD,
 					" => dimming screen in order to disable touch events");
 			params.screenBrightness = 0.1f;
 			view.setVisibility(View.INVISIBLE);
 			showNavigationBar(false);
 			window.addFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN);
 		} else {
-			Lg.i(LOGTAG, "proximity sensors distance=", Float.valueOf(distance), " above threshold=", Float.valueOf(PROXIMITY_DISTANCE_THRESHOLD),
+			Lg.i(LOGTAG, "proximity sensors distance=", distance, " above threshold=", PROXIMITY_DISTANCE_THRESHOLD,
 					" => enabling touch events (no screen dimming)");
 			params.screenBrightness = WindowManager.LayoutParams.BRIGHTNESS_OVERRIDE_NONE;
 			view.setVisibility(View.VISIBLE);
